+++ server/TWS Foundation/Server.Quality/Suit/Controllers/Business/Q_DriversController.cs
@@ -177,7 +177,25 @@ public class Q_DriversController
             ]);
             #endregion
         }
+      
         #endregion
+    }
+    [Fact]
+    public async Task Delete() {
+        string testTag = Guid.NewGuid().ToString()[..3];
+        List<Driver> mock = [MockFactory(testTag)];
+
+        // Create a new record to delete.
+        (HttpStatusCode Status, GenericFrame response) = await Post("Create", mock, true);
+        SetBatchOut<Driver> estela = Framing<SuccessFrame<SetBatchOut<Driver>>>(response).Estela;
+        Assert.Equal(HttpStatusCode.OK, Status);
+        Assert.Empty(estela.Failures);
+
+        // Deleting the previous record.
+        Driver newAccount = estela.Successes.First();
+        (HttpStatusCode DeleteStatus, _) = await Post("Delete", newAccount, true);
+        Assert.Equal(HttpStatusCode.OK, DeleteStatus);
+
     }
 
 }