+++ server/TWS Foundation/TWS Customer/Services/Interfaces/IDriversExternalsService.cs
@@ -3,9 +3,13 @@ using CSM_Foundation.Database.Models.Out;
 
 using TWS_Business.Sets;
 
+using TWS_Security.Sets;
+
 namespace TWS_Customer.Services.Interfaces;
 public interface IDriversExternalsService {
     Task<SetViewOut<DriverExternal>> View(SetViewOptions<DriverExternal> Options);
     Task<SetBatchOut<DriverExternal>> Create(DriverExternal[] Driver);
     Task<RecordUpdateOut<DriverExternal>> Update(DriverExternal Driver);
+    Task<DriverExternal> Delete(DriverExternal Driver);
+
 }